+++ src/Dephpug/Config.php
@@ -16,7 +16,8 @@ class Config
         ],
         'debugger' => [
             'port' => 9005,
-            'host' => 'localhost'
+            'host' => 'localhost',
+            'lineOffset' => 6
         ],
         'options' => [
             'verboseMode' => false